+++ src/NeatMapper.Async/AsyncMapperExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using NeatMapper.Async.Internal;
+using NeatMapper.Common;
 using System;
 using System.Collections.Generic;
 using System.Threading;
@@ -32,7 +33,48 @@ namespace NeatMapper.Async {
 				throw new ArgumentNullException(nameof(mapper));
 			if (source == null)
 				throw new ArgumentNullException(nameof(source));
-			return TaskUtils.AwaitTask<TDestination>(mapper.MapAsync(source, source.GetType(), typeof(TDestination), cancellationToken));
+			return TaskUtils.AwaitTask<TDestination>(mapper.MapAsync(source, source.GetType(), typeof(TDestination), null, cancellationToken));
+
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+#nullable enable
+#endif
+		}
+
+		/// <summary>
+		/// Maps an object to a new one asynchronously
+		/// </summary>
+		/// <typeparam name="TDestination">type of the destination object to create, used to retrieve the available maps</typeparam>
+		/// <param name="source">
+		/// object to map, may NOT be null as the source type will be retrieved from it,
+		/// which will be used to retrieve the available maps
+		/// </param>
+		/// <param name="mappingOptions">additional options for the current map, null to use default ones</param>
+		/// <param name="cancellationToken">cancellation token used to cancel async operations, will be forwarded to all the contexts in the mapping</param>
+		/// <returns>a task which when completed returns the newly created object, which may be null</returns>
+		public static Task<
+#if NET5_0_OR_GREATER
+			TDestination?
+#else
+			TDestination
+#endif
+			> MapAsync<TDestination>(this IAsyncMapper mapper,
+			object source,
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			AsyncMappingOptions?
+#else
+			AsyncMappingOptions
+#endif
+			mappingOptions, CancellationToken cancellationToken = default) {
+
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+#nullable disable
+#endif
+
+			if (mapper == null)
+				throw new ArgumentNullException(nameof(mapper));
+			if (source == null)
+				throw new ArgumentNullException(nameof(source));
+			return TaskUtils.AwaitTask<TDestination>(mapper.MapAsync(source, source.GetType(), typeof(TDestination), mappingOptions, cancellationToken));
 
 #if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
 #nullable enable
@@ -68,7 +110,50 @@ namespace NeatMapper.Async {
 
 			if (mapper == null)
 				throw new ArgumentNullException(nameof(mapper));
-			return TaskUtils.AwaitTask<TDestination>(mapper.MapAsync(source, typeof(TSource), typeof(TDestination), cancellationToken));
+			return TaskUtils.AwaitTask<TDestination>(mapper.MapAsync(source, typeof(TSource), typeof(TDestination), null, cancellationToken));
+
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+#nullable enable
+#endif
+		}
+
+		/// <summary>
+		/// Maps an object to a new one asynchronously
+		/// </summary>
+		/// <typeparam name="TSource">type of the object to map, used to retrieve the available maps</typeparam>
+		/// <typeparam name="TDestination">type of the destination object to create, used to retrieve the available maps</typeparam>
+		/// <param name="source">object to map, may be null</param>
+		/// <param name="mappingOptions">additional options for the current map, null to use default ones</param>
+		/// <param name="cancellationToken">cancellation token used to cancel async operations, will be forwarded to all the contexts in the mapping</param>
+		/// <returns>a task which when completed returns the newly created object, which may be null</returns>
+		public static Task<
+#if NET5_0_OR_GREATER
+			TDestination?
+#else
+			TDestination
+#endif
+			> MapAsync<TSource, TDestination>(this IAsyncMapper mapper,
+#if NET5_0_OR_GREATER
+			TSource?
+#else
+			TSource
+#endif
+			source,
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			AsyncMappingOptions?
+#else
+			AsyncMappingOptions
+#endif
+			mappingOptions,
+			CancellationToken cancellationToken = default) {
+
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+#nullable disable
+#endif
+
+			if (mapper == null)
+				throw new ArgumentNullException(nameof(mapper));
+			return TaskUtils.AwaitTask<TDestination>(mapper.MapAsync(source, typeof(TSource), typeof(TDestination), mappingOptions, cancellationToken));
 
 #if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
 #nullable enable
@@ -128,7 +213,7 @@ namespace NeatMapper.Async {
 		/// <typeparam name="TDestination">type of the destination object, used to retrieve the available maps</typeparam>
 		/// <param name="source">object to be mapped, may be null</param>
 		/// <param name="destination">object to map to, may be null</param>
-		/// <param name="mappingOptions">additional options for the current map</param>
+		/// <param name="mappingOptions">additional options for the current map, null to use default ones</param>
 		/// <param name="cancellationToken">cancellation token used to cancel async operations, will be forwarded to all the contexts in the mapping</param>
 		/// <returns>
 		/// a task which when completed returns the resulting object of the mapping, which can be <paramref name="destination"/>
@@ -154,9 +239,9 @@ namespace NeatMapper.Async {
 #endif
 			destination,
 #if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
-			MappingOptions?
+			AsyncMappingOptions?
 #else
-			MappingOptions
+			AsyncMappingOptions
 #endif
 			mappingOptions,
 			CancellationToken cancellationToken = default) {
@@ -208,12 +293,7 @@ namespace NeatMapper.Async {
 			ICollection<TDestinationElement>
 #endif
 			destination,
-#if NET5_0_OR_GREATER
-			Func<TSourceElement?, TDestinationElement?, MatchingContext, bool>
-#else
-			Func<TSourceElement, TDestinationElement, MatchingContext, bool>
-#endif
-			matcher,
+			MatchMapDelegate<TSourceElement, TDestinationElement> matcher,
 			CancellationToken cancellationToken = default) {
 
 #if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
@@ -229,7 +309,7 @@ namespace NeatMapper.Async {
 				typeof(IEnumerable<TSourceElement>),
 				destination,
 				typeof(ICollection<TDestinationElement>),
-				new MappingOptions {
+				new AsyncMappingOptions {
 					Matcher = (s, d, c) => (s is TSourceElement || object.Equals(s, default(TSourceElement))) &&
 						(d is TDestinationElement || object.Equals(d, default(TDestinationElement))) &&
 						matcher((TSourceElement)s, (TDestinationElement)d, c)
@@ -248,7 +328,7 @@ namespace NeatMapper.Async {
 		/// <param name="source">collection to be mapped, may be null</param>
 		/// <param name="destination">collection to map to, may be null</param>
 		/// <param name="matcher">matching method to be used to match elements of the <paramref name="source"/> and <paramref name="destination"/> collections</param>
-		/// <param name="removeNotMatchedDestinationElements">if true will remove all the elements from <paramref name="destination"/> which do not have a corresponding element in <paramref name="source"/></param>
+		/// <param name="removeNotMatchedDestinationElements">if true will remove all the elements from <paramref name="destination"/> which do not have a corresponding element in <paramref name="source"/>, null to use default setting</param>
 		/// <param name="cancellationToken">cancellation token used to cancel async operations, will be forwarded to all the contexts in the mapping</param>
 		/// <returns>
 		/// a task which when completed returns the resulting collection of the mapping, which can be <paramref name="destination"/>
@@ -275,12 +355,7 @@ namespace NeatMapper.Async {
 			ICollection<TDestinationElement>
 #endif
 			destination,
-#if NET5_0_OR_GREATER
-			Func<TSourceElement?, TDestinationElement?, MatchingContext, bool>
-#else
-			Func<TSourceElement, TDestinationElement, MatchingContext, bool>
-#endif
-			matcher,
+			MatchMapDelegate<TSourceElement, TDestinationElement> matcher,
 			bool? removeNotMatchedDestinationElements,
 			CancellationToken cancellationToken = default) {
 
@@ -297,7 +372,7 @@ namespace NeatMapper.Async {
 				typeof(IEnumerable<TSourceElement>),
 				destination,
 				typeof(ICollection<TDestinationElement>),
-				new MappingOptions {
+				new AsyncMappingOptions {
 					Matcher = (s, d, c) => (s is TSourceElement || object.Equals(s, default(TSourceElement))) &&
 						(d is TDestinationElement || object.Equals(d, default(TDestinationElement))) &&
 						matcher((TSourceElement)s, (TDestinationElement)d, c),