+++ src/NeatMapper.Async/AsyncMappingOptions.cs
@@ -1,0 +1,8 @@
+﻿namespace NeatMapper.Async {
+	/// <summary>
+	/// Mapping options to use for the mapping operations, this will override any configuration options defined in <see cref="Configuration.MapperConfigurationOptions"/>
+	/// </summary>
+	public sealed class AsyncMappingOptions : MappingOptions {
+
+	}
+}