+++ src/NeatMapper.Async/IAsyncMapper.cs
@@ -14,6 +14,7 @@ namespace NeatMapper.Async {
 		/// <param name="source">object to map, may be null</param>
 		/// <param name="sourceType">type of the object to map, used to retrieve the available maps</param>
 		/// <param name="destinationType">type of the destination object to create, used to retrieve the available maps</param>
+		/// <param name="mappingOptions">additional options for the current map, null to use default ones</param>
 		/// <param name="cancellationToken">cancellation token used to cancel async operations, will be forwarded to all the contexts in the mapping</param>
 		/// <returns>a task which when completed returns the newly created object of <paramref name="destinationType"/>, which may be null</returns>
 		Task<
@@ -31,6 +32,12 @@ namespace NeatMapper.Async {
 			source,
 			Type sourceType,
 			Type destinationType,
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			AsyncMappingOptions?
+#else
+			AsyncMappingOptions
+#endif
+			mappingOptions = null,
 			CancellationToken cancellationToken = default);
 
 		/// <summary>
@@ -42,7 +49,7 @@ namespace NeatMapper.Async {
 		/// <param name="sourceType">type of the object to be mapped, used to retrieve the available maps</param>
 		/// <param name="destination">object to map to, may be null</param>
 		/// <param name="destinationType">type of the destination object, used to retrieve the available maps</param>
-		/// <param name="mappingOptions">additional options for the current map</param>
+		/// <param name="mappingOptions">additional options for the current map, null to use default ones</param>
 		/// <param name="cancellationToken">cancellation token used to cancel async operations, will be forwarded to all the contexts in the mapping</param>
 		/// <returns>
 		/// a task which when completed returns the resulting object of the mapping of <paramref name="destinationType"/> type,
@@ -70,9 +77,9 @@ namespace NeatMapper.Async {
 			destination,
 			Type destinationType,
 #if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
-			MappingOptions?
+			AsyncMappingOptions?
 #else
-			MappingOptions
+			AsyncMappingOptions
 #endif
 			mappingOptions = null,
 			CancellationToken cancellationToken = default);